+++ conf/buster-test-integration.js
@@ -1,0 +1,7 @@
+module.exports.unit = {
+  environment: 'node',
+  rootPath: process.cwd(),
+  tests: [
+    'test-integration/**/*.js'
+  ]
+};